Rename the main directory to pZudoku and fix imports
