+++ pZudoku/api.py
@@ -1,0 +1,1 @@
+from pZudoku.hb6d_board import HB6DBoard as Board
